% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignEnzyme.R
\name{assignEnzyme}
\alias{assignEnzyme}
\title{assignEnzyme}
\usage{
assignEnzyme(filename)
}
\arguments{
\item{x1:}{\code{filename}}
}
\value{
y1: enzymename
}
\description{
This function extract the enzyme name from
a filename
}

