% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignPrep.R
\name{assignPrep}
\alias{assignPrep}
\title{assignPrep}
\usage{
assignPrep(filename)
}
\arguments{
\item{x1:}{\code{filename}}
}
\value{
y1: enzymename
}
\description{
This function extract the preparation method from
a filename
}

