% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignTemplate.R
\name{assignTemplate}
\alias{assignTemplate}
\title{assignTemplate}
\usage{
assignTemplate(filename)
}
\arguments{
\item{x1:}{\code{filename}}
}
\value{
y1: template
}
\description{
This function extract the tempate (DNA/RNA) from
a filename
}

